+++ examples/resources/upcloud_storage_backup/resource.tf
@@ -1,0 +1,4 @@
+resource "upcloud_storage_backup" "backup_1" {
+  title          = "backup"
+  source_storage = "01185ec5-1b0b-4cbc-a968-eb920ac7572d"
+}